requirements.txt
# Core dependencies
streamlit>=1.31.0
openai>=1.10.0
python-dotenv>=1.0.0

# LangChain dependencies (for Week 3+)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Optional: for RAG and document processing
# pypdf>=3.17.0
# faiss-cpu>=1.7.4
# chromadb>=0.4.22

# Optional: for data analysis
# pandas>=2.0.0
# numpy>=1.24.0

# Optional: for voice features
# whisper>=1.1.10
# gtts>=2.5.0
